auto clear cache logic
