+++ docs/demo/auto-clear.md
@@ -1,0 +1,3 @@
+## Auto Clear
+
+<code src="../examples/auto-clear.tsx">
