+++ src/CacheContext.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import CacheEntity from './Cache';
 
 const CacheContext = React.createContext({
-  autoClean: true,
-  cache: new CacheEntity<any, [number, any]>(),
+  autoClear: false,
+  cache: new CacheEntity<[number, any]>(),
 });
 
 export default CacheContext;
